Move normalization of gravity to the format helper
