+++ js/composites/HydrometerTemperature.js
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 
 import { Gravity, GravityHelp, Temperature, TemperatureHelp } from '../components';
 import { FormGroup, StaticFormGroup } from './form';
-import { correctHydrometerTemperature, normalizeGravity } from '../calculation';
-import { formatAsFloat } from "../helpers/format";
+import { correctHydrometerTemperature } from '../calculation';
+import { formatAsFloat, normalizeGravity } from "../helpers/format";
 import { validateTemperature, validateGravity } from "../validations";
 
 const calculate = ( gravity, temperature, calibration ) => {
